+++ CalendarSchedule.API/Controllers/UserController.cs
@@ -15,8 +15,7 @@ public class UserController(IUserService _userService) : ControllerBase
 {
     [HttpGet]
     [Route("All")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UserDto[]))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(PagedResult<UserDto>))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetAll([FromQuery] int page = 1, [FromQuery] int size = 10, [FromQuery] string search = "")
     {
@@ -24,32 +23,14 @@ public class UserController(IUserService _userService) : ControllerBase
         if (usersDto.IsFailure)
             return NotFound(usersDto.Error);
 
-        var totalUsers = await _userService.TotalUsers(search);
-        if (totalUsers.IsFailure)
-            return NotFound(totalUsers.Error);
-
-        decimal totalData = totalUsers.Value;
-        decimal totalPage = (totalData / size) <= 0 ? 1 : Math.Ceiling((totalData / size));
-
-        if (size == 1)
-            totalPage = totalData;
-
         var users = usersDto.Value;
 
-        return Ok(new
-        {
-            totalData,
-            page,
-            totalPage,
-            size,
-            users
-        });
-
+        return Ok(users);
     }
 
     [HttpGet]
     [Route("ManagerAll")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UserDto[]))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(PagedResult<UserDto>))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetManagerAll([FromQuery] int page = 1, [FromQuery] int size = 10, [FromQuery] string search = "")
@@ -58,32 +39,14 @@ public class UserController(IUserService _userService) : ControllerBase
         if (usersDto.IsFailure)
             return NotFound(usersDto.Error);
 
-        var totalUsers = await _userService.TotalUsers(search);
-        if (totalUsers.IsFailure)
-            return NotFound(totalUsers.Error);
-
-        decimal totalData = totalUsers.Value;
-        decimal totalPage = (totalData / size) <= 0 ? 1 : Math.Ceiling((totalData / size));
-
-        if (size == 1)
-            totalPage = totalData;
-
         var users = usersDto.Value;
 
-        return Ok(new
-        {
-            totalData,
-            page,
-            totalPage,
-            size,
-            users
-        });
-
+        return Ok(users);
     }
 
     [HttpGet]
     [Route("ManagerAllByUserCurrent")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UserDto[]))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(PagedResult<UserDto>))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetManagerAllByUserCurrent([FromQuery] int page = 1, [FromQuery] int size = 10, [FromQuery] string search = "", [FromQuery] string username = "")
@@ -92,26 +55,9 @@ public class UserController(IUserService _userService) : ControllerBase
         if (usersDto.IsFailure)
             return NotFound(usersDto.Error);
 
-        var totalUsers = await _userService.TotalUsers(search);
-        if (totalUsers.IsFailure)
-            return NotFound(totalUsers.Error);
-
-        decimal totalData = totalUsers.Value;
-        decimal totalPage = (totalData / size) <= 0 ? 1 : Math.Ceiling((totalData / size));
-
-        if (size == 1)
-            totalPage = totalData;
-
         var users = usersDto.Value;
 
-        return Ok(new
-        {
-            totalData,
-            page,
-            totalPage,
-            size,
-            users
-        });
+        return Ok(users);
     }
 
     [HttpGet("{id:int}", Name = "GetUser")]
@@ -120,6 +66,12 @@ public class UserController(IUserService _userService) : ControllerBase
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetById(int id)
     {
+        if (id <= 0)
+        {
+            var error = Result.Failure(Error.BadRequest($"Id ({id}) inválido"));
+            return BadRequest(error);
+        }
+
         var userDto = await _userService.GetById(id);
         if (userDto.IsFailure)
             return NotFound(userDto.Error);
@@ -132,6 +84,12 @@ public class UserController(IUserService _userService) : ControllerBase
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetManagerUsername(string username)
     {
+        if (string.IsNullOrEmpty(username))
+        {
+            var error = Result.Failure(Error.BadRequest($"Username nulo"));
+            return BadRequest(error);
+        }
+
         var userDto = await _userService.GetManagerUsername(username);
         if (userDto.IsFailure)
             return NotFound(userDto.Error);
@@ -145,11 +103,28 @@ public class UserController(IUserService _userService) : ControllerBase
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(Error))]
     public async Task<IActionResult> Post([FromBody] UserDto userDto)
     {
+        if (!ModelState.IsValid)
+        {
+            var errorMessage = ErroMoldeState();
+
+            var error = Result.Failure(Error.BadRequest($"Erro de validação no objeto ({nameof(UserDto)}): {errorMessage}"));
+            return BadRequest(error);
+        }
+
         var user = await _userService.Create(userDto);
         if (user.IsFailure)
-            return NotFound(user.Error);
+        {
+            return user.Error.Code switch
+            {
+                "NotFound" => NotFound(user.Error),
+                "BadRequest" => BadRequest(user.Error),
+                "Internal" => StatusCode(StatusCodes.Status500InternalServerError, user.Error),
+                _ => StatusCode(StatusCodes.Status500InternalServerError, user.Error) // fallback para erro desconhecido
+            };
+        }
 
         var createResult = user.Value;
 
@@ -162,14 +137,36 @@ public class UserController(IUserService _userService) : ControllerBase
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(Error))]
     public async Task<IActionResult> Put(int id, [FromBody] UserDto userDto)
     {
         if (id <= 0)
-            return BadRequest("Id inválidos");
-
+        {
+            var error = Result.Failure(Error.BadRequest($"Id ({id}) inválido"));
+            return BadRequest(error);
+        }
+        if (!ModelState.IsValid)
+        {
+            var errorMessage = ErroMoldeState();
+            var error = Result.Failure(Error.BadRequest($"Erro de validação do objeto ({nameof(UserDto)}): {errorMessage}"));
+            return BadRequest(error);
+        }
+        if (userDto.Id != id)
+        {
+            var error = Result.Failure(Error.BadRequest($"Id ({id}) inválido"));
+            return BadRequest(error);
+        }
         var user = await _userService.Update(userDto);
         if (user.IsFailure)
-            return NotFound(user.Error);
+        {
+            return user.Error.Code switch
+            {
+                "NotFound" => NotFound(user.Error),
+                "BadRequest" => BadRequest(user.Error),
+                "Internal" => StatusCode(StatusCodes.Status500InternalServerError, user.Error),
+                _ => StatusCode(StatusCodes.Status500InternalServerError, user.Error) // fallback para erro desconhecido
+            };
+        }
 
         return Ok(user.Value);
     }
@@ -180,16 +177,34 @@ public class UserController(IUserService _userService) : ControllerBase
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(Error))]
     public async Task<ActionResult<UserDto>> Delete(int id)
     {
+        if (id <= 0)
+        {
+            var error = Result.Failure(Error.BadRequest($"Id inválido: {id}"));
+            return BadRequest(error);
+        }
+
         var result = await _userService.Delete(id);
         if (result.IsFailure)
-            return NotFound(result.Error);
+        {
+            return result.Error.Code switch
+            {
+                "NotFound" => NotFound(result.Error),
+                "BadRequest" => BadRequest(result.Error),
+                "Internal" => StatusCode(StatusCodes.Status500InternalServerError, result.Error),
+                _ => StatusCode(StatusCodes.Status500InternalServerError, result.Error) // fallback para erro desconhecido
+            };
+        }
 
-        var userDto = await _userService.GetById(id);
-        if (userDto.IsFailure)
-            return NotFound(userDto.Error);
-
-        return Ok(userDto.Value);
+        return Ok(result.Value);
+    }
+    private string ErroMoldeState()
+    {
+        var errorMessage = string.Join("; ", ModelState.Values
+                                              .SelectMany(x => x.Errors)
+                                              .Select(x => x.ErrorMessage));
+        return errorMessage;
     }
 }
